+++ ios/DBModule.swift
@@ -1,6 +1,6 @@
 //
 //  DBModule.swift
-//  ScrabbleHelper
+//  Wyrazowo
 //
 //  Created by Adam Kruczek on 03/04/2023.
 //