+++ ios/RCTDBModule.h
@@ -1,6 +1,6 @@
 //
 //  RCTDBModule.h
-//  ScrabbleHelper
+//  Wyrazowo
 //
 //  Created by Adam Kruczek on 03/04/2023.
 //