+++ ios/RCTDBModule.m
@@ -1,6 +1,6 @@
 //
 //  RCTDBModule.m
-//  ScrabbleHelper
+//  Wyrazowo
 //
 //  Created by Adam Kruczek on 03/04/2023.
 //