+++ src/charade/components/charade-field/charade-field.tsx
@@ -5,17 +5,18 @@ interface Props {
   count: number;
   isActive: boolean;
   isSent: boolean;
+  isError: boolean;
   word: string;
   index: number;
   content: string;
   onPress: () => void;
 }
 
-export const CharadeField = ({ count, onPress, content, isSent, index, word, isActive }: Props) => {
+export const CharadeField = ({ count, onPress, content, isSent, isError, index, word, isActive }: Props) => {
   return (
-    <CharadeFieldContainer {...{ onPress, index, count, content, word, isSent }}>
-      <CharadeFieldInnerContainer isActive={isActive} />
-      <CharadeFieldContentText children={content} />
+    <CharadeFieldContainer {...{ onPress, index, count, content, word, isError, isSent }}>
+      <CharadeFieldInnerContainer {...{ isError, isActive }} />
+      <CharadeFieldContentText isError={isError} children={content} />
     </CharadeFieldContainer>
   )
 }